campuscore/requirements.txt
django>=5.1,<5.2
whitenoise[brotli]>=6,<7
pyyaml
djangorestframework
djangorestframework-simplejwt
django-cors-headers
black
isort
